+++ src/Calendar.less
@@ -6,13 +6,10 @@
  * All rights reserved.
  */
 
-
-@calendarFullCls: kuma-full-calendar;
 @__calendarPrefixCls: kuma-calendar;
 @__calendarTimePickerPrefixCls: kuma-time-picker;
 @__selectPrefixCls: kuma-select2;
 @calendar-header-hover-bg: hsv(round(hsvhue(@brand-primary)), round(hsvsaturation(@brand-primary)) - 17%, round(hsvvalue(@brand-primary)) + 1%); // B1-1（s-17,b+1）
-
 @import "style/common/index";
 @import "style/panel/index";
 @import "style/timepicker/index";
