+++ src/fullTime/TimeTable.jsx
@@ -2,6 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import TimeTHead from './TimeTHead';
 import TimeTBody from './TimeTBody';
+
 class TimeTable extends React.Component {
   constructor(props) {
     super(props);
@@ -10,34 +11,43 @@ class TimeTable extends React.Component {
       slicePiece: 0,
     };
   }
-  getTimeLine() {
-    let { startHour, gapMinute, endHour } = this.props;
-    startHour = startHour ? parseInt(startHour) : 9;
-    endHour = endHour ? parseInt(endHour) : 23;
-    gapMinute = gapMinute ? parseInt(gapMinute) : 60;
-    if (endHour < startHour) {
-      endHour = startHour;
-    }
-    let slicePiece = Math.ceil(((endHour - startHour) * 60) / gapMinute);
-    this.setState({ slicePiece });
-  }
+
   componentDidMount() {
     this.getTimeLine();
   }
+
+  getTimeLine() {
+    let { startHour, gapMinute, endHour } = this.props;
+    startHour = startHour ? parseInt(startHour, 10) : 9;
+    endHour = endHour ? parseInt(endHour, 10) : 23;
+    gapMinute = gapMinute ? parseInt(gapMinute, 10) : 60;
+    if (endHour < startHour) {
+      endHour = startHour;
+    }
+    const slicePiece = Math.ceil(((endHour - startHour) * 60) / gapMinute);
+    this.setState({ slicePiece });
+  }
+
   renderEvents() {
-    let { scheduleRender, startHour, gapMinute, endHour, value, type } = this.props;
+    const {
+      scheduleRender, startHour, gapMinute, endHour, value, type,
+    } = this.props;
     const { slicePiece } = this.state;
-    let renderOpts = { startHour, gapMinute, endHour, type, slicePiece, current: value };
+    const renderOpts = {
+      startHour, gapMinute, endHour, type, slicePiece, current: value,
+    };
     if (scheduleRender) {
       const content = scheduleRender(renderOpts);
       return <div className="events-wrapper">{content}</div>;
     }
+    return null;
   }
+
   render() {
     const { slicePiece } = this.state;
-    const { showCount, prefixCls, timeRender } = this.props;
-    let timeCls = `${prefixCls}-time`;
-    let tableCls = `${prefixCls}-table`;
+    const { showCount, prefixCls } = this.props;
+    const timeCls = `${prefixCls}-time`;
+    const tableCls = `${prefixCls}-table`;
     return (
       <div className={timeCls}>
         <table className={tableCls} cellSpacing="0" role="grid">
