+++ src/fullWeek/WeekTable.jsx
@@ -2,6 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import WeekTHead from './WeekTHead';
 import WeekTBody from './WeekTBody';
+
 class WeekTable extends React.Component {
   constructor(props) {
     super(props);
@@ -10,41 +11,48 @@ class WeekTable extends React.Component {
       slicePiece: 0,
     };
   }
-  getTimeLine() {
-    let { startHour, gapMinute, endHour } = this.props;
-    startHour = startHour ? parseInt(startHour) : 9;
-    endHour = endHour ? parseInt(endHour) : 23;
-    gapMinute = gapMinute ? parseInt(gapMinute) : 60;
-    if (endHour < startHour) {
-      endHour = startHour;
-    }
-    let slicePiece = Math.ceil(((endHour - startHour) * 60) / gapMinute);
-    this.setState({ slicePiece });
-  }
+
   componentDidMount() {
     this.getTimeLine();
   }
+
+  getTimeLine() {
+    let { startHour, gapMinute, endHour } = this.props;
+    startHour = startHour ? parseInt(startHour, 10) : 9;
+    endHour = endHour ? parseInt(endHour, 10) : 23;
+    gapMinute = gapMinute ? parseInt(gapMinute, 10) : 60;
+    if (endHour < startHour) {
+      endHour = startHour;
+    }
+    const slicePiece = Math.ceil(((endHour - startHour) * 60) / gapMinute);
+    this.setState({ slicePiece });
+  }
+
   renderEvents() {
-    let { scheduleRender, startHour, gapMinute, endHour, value, type } = this.props;
+    const {
+      scheduleRender, startHour, gapMinute, endHour, value, type,
+    } = this.props;
 
     const { slicePiece } = this.state;
-    let renderOpts = { startHour, gapMinute, endHour, slicePiece, type, current: value };
+    const renderOpts = {
+      startHour, gapMinute, endHour, slicePiece, type, current: value,
+    };
     if (scheduleRender) {
       const content = scheduleRender(renderOpts);
-
       return <div className="events-wrapper">{content}</div>;
     }
+    return null;
   }
+
   render() {
     const { slicePiece } = this.state;
-    const props = this.props;
-    const prefixCls = props.prefixCls;
-    let weekCls = `${prefixCls}-date`;
+    const { prefixCls } = this.props;
+    const weekCls = `${prefixCls}-date`;
     return (
       <div className={weekCls}>
         <table className={`${prefixCls}-table`} cellSpacing="0" role="grid">
-          <WeekTHead {...props} />
-          <WeekTBody {...props} slicePiece={slicePiece} />
+          <WeekTHead {...this.props} />
+          <WeekTBody {...this.props} slicePiece={slicePiece} />
         </table>
         {this.renderEvents()}
       </div>
