+++ src/style/panel/Week.less
@@ -1,4 +1,4 @@
-.@{calendarFullCls}-week-panel {
+.@{__calendarPrefixCls}-full-week-panel {
   &-column-header {
     height: 32px;
     padding: 8px;
@@ -13,7 +13,7 @@
     .header-date {
       padding-right: 5px;
     }
-    &.@{calendarFullCls}-week-panel-today {
+    &.@{__calendarPrefixCls}-full-week-panel-today {
       color: @brand-primary;
     }
   }
@@ -25,7 +25,7 @@
     font-size: 12px;
     border-top: 1px solid #e8ebee;
     color: rgba(0, 0, 0, 0.6);
-    .@{calendarFullCls}-week-panel-timeline-cell {
+    .@{__calendarPrefixCls}-full-week-panel-timeline-cell {
       height: 48px;
       line-height: 48px;
       position: relative;
