+++ src/style/panel/index.less
@@ -2,7 +2,7 @@
 @import "Calendar";
 @import "Time";
 @import "TimePanel";
-@import "TimeBody";
+@import "FullTime";
 @import "MonthPanel";
 @import "YearPanel";
 @import "DecadePanel";
